requirements.txt
fastapi==0.115.8
uvicorn==0.30.6
python-dotenv==1.0.1
psutil==6.1.0
pytest==8.3.4
httpx==0.27.2
python-multipart==0.0.9
